# [Cluster Test] Add a performance test suite for state syncing fullnodes.