--- a/testsuite/cluster-test/src/experiments/mod.rs
+++ b/testsuite/cluster-test/src/experiments/mod.rs
@@ -13,6 +13,7 @@ mod reboot_cluster;
 mod reboot_random_validators;
 mod reconfiguration_test;
 mod recovery_time;
+mod state_sync_performance;
 mod twin_validator;
 mod versioning_test;
 
@@ -35,6 +36,7 @@ pub use reboot_cluster::{RebootCluster, RebootClusterParams};
 pub use reboot_random_validators::{RebootRandomValidators, RebootRandomValidatorsParams};
 pub use reconfiguration_test::{Reconfiguration, ReconfigurationParams};
 pub use recovery_time::{RecoveryTime, RecoveryTimeParams};
+pub use state_sync_performance::{StateSyncPerformance, StateSyncPerformanceParams};
 pub use twin_validator::{TwinValidators, TwinValidatorsParams};
 pub use versioning_test::{ValidatorVersioning, ValidatorVersioningParams};
 
@@ -156,6 +158,7 @@ pub fn get_experiment(name: &str, args: &[String], cluster: &Cluster) -> Box<dyn
     known_experiments.insert("reboot_cluster", f::<RebootClusterParams>());
     known_experiments.insert("reconfiguration", f::<ReconfigurationParams>());
     known_experiments.insert("load_test", f::<LoadTestParams>());
+    known_experiments.insert("state_sync_performance", f::<StateSyncPerformanceParams>());
 
     let builder = known_experiments.get(name).expect("Experiment not found");
     builder(args, cluster)