## Motivation

This PR adds a new state sync performance benchmark to cluster test (to measure the performance of state syncing fullnodes). For now, the benchmark is somewhat primitive (i.e., it simply measures how long a fullnode takes to synchronize state from version 0 to the expected version of the validators around it). Current results are showing around 1.7k tx/s in cluster test (see the benchmark output below).

Once this benchmark lands, I'll follow up by enabling the benchmark to run on all future PRs and display the results. This will allow us to keep a closer eye on state sync performance going forward.

### Have you read the [Contributing Guidelines on pull requests](https://github.com/diem/diem/blob/main/CONTRIBUTING.md#pull-requests)?

Yes.

## Test Plan

I've run this benchmark several times manually using `./scripts/cti ... --suite state_sync`. The output is as follows:

```
...
2021-04-07T21:16:51.355804Z [main] INFO testsuite/cluster-test/src/experiments/state_sync_performance.rs:93 The validator is now synced at version: 58806
2021-04-07T21:16:51.355817Z [main] INFO testsuite/cluster-test/src/experiments/state_sync_performance.rs:99 Waiting for the fullnode to wake up: fn-14-0(192.168.239.79)
2021-04-07T21:17:07.278455Z [main] INFO testsuite/cluster-test/src/experiments/state_sync_performance.rs:109 The fullnode is now up. Waiting for it to state sync to the expected version: 58806
2021-04-07T21:17:40.561229Z [main] INFO testsuite/cluster-test/src/experiments/state_sync_performance.rs:117 The fullnode has caught up to version: 58806
2021-04-07T21:17:40.561243Z [main] INFO testsuite/cluster-test/src/experiments/state_sync_performance.rs:130 Time to state sync 33.282774558s
2021-04-07T21:17:40.561245Z [main] INFO testsuite/cluster-test/src/experiments/state_sync_performance.rs:133 State sync throughput : 1766.9010275824771 txn/sec
2021-04-07T21:17:40.561266Z [main] INFO testsuite/cluster-test/src/main.rs:686 ^[[1mExperiment finished, waiting until all affected validators recover^[[m
2021-04-07T21:17:40.561268Z [main] INFO testsuite/cluster-test/src/main.rs:749 Waiting for all nodes to be healthy
2021-04-07T21:17:45.592993Z [main] INFO testsuite/cluster-test/src/main.rs:770 All nodes are now healthy. Checking json rpc endpoints of validators and full nodes
2021-04-07T21:17:45.956058Z [main] INFO testsuite/cluster-test/src/main.rs:795 All json rpc endpoints are healthy
2021-04-07T21:17:45.956072Z [main] INFO testsuite/cluster-test/src/main.rs:694 Experiment completed
2021-04-07T21:17:45.956239Z [main] INFO testsuite/cluster-test/src/main.rs:611 Suite completed in 153.545571596s
2021-04-07T21:17:45.956265Z [main] INFO testsuite/cluster-test/src/main.rs:622 
====json-report-begin===
{
  "metrics": [
    {
      "experiment": "StateSyncPerformance",
      "metric": "state_sync_throughput",
      "value": 1766.9010275824771
    }
  ],
  "text": "Test runner setup time spent 249 secs\nState sync throughput : 1766.9010275824771 txn/sec, time spent 153 secs"
}
```


## Related PRs

None.

## If targeting a release branch, please fill the below out as well

None.