@aching 2021-04-09
Btw, it might be interesting to get in a run before and after the perf wins for state sync v1.1 to see how it fairs in this environment vs the laptop.